+++ XEngine/src/math.h
@@ -26,7 +26,7 @@ static float clamp(float value, float min, float max)
 	return ((std::min)(((std::max)(value, min)), max));
 }
 
-static float clamp(int value, int min, int max)
+static int clamp(int value, int min, int max)
 {
 	return ((std::min)(((std::max)(value, min)), max));
 }
